test: added tests
