+++ event-sourcing-cqrs/package.json
@@ -8,7 +8,7 @@
     "prestart": "yarn build",
     "start": "node .",
     "lint": "tslint -c tslint.json -p tsconfig.json --fix",
-    "test": "mocha -r ts-node/register --exit test/**/*.test.ts"
+    "test": "./bin/db-start && mocha -r ts-node/register --exit test/**/*.test.ts"
   },
   "dependencies": {
     "dotenv": "^8.2.0",
