+++ event-sourcing-cqrs/tsconfig.json
@@ -15,6 +15,7 @@
       }
   },
   "include": [
-      "src/*"
+      "src/*",
+      "test/*",
   ]
 }
